requirements.txt
bandit==1.8.3
black==25.1.0
fastapi==0.115.12
fastapi-cache2[redis]
flake8==7.1.2
isort==6.0.1
loguru==0.7.3
mypy==1.15.0
pre-commit==4.2.0
pydantic
python-dotenv==1.0.1
redis==4.6.0
redis-om==0.4.7
uvicorn==0.34.0
